Added etcd comms channel
The AgentCommunicator object has been broken up into a
hierarchy, I've added an etcd version (which, note, lacks a
journal as yet, and needs one), and made it so that the etcd
one runs.  The agent doesn't use the etcd one right now, so there
is a simple bit of debug code that fires off both new and old code
for the moment.  The etcd code will also pick up bind notifications
and I've left a bit of unrefined code in there for that purpose
(which you can trip with etcdctl).

Change-Id: I82681d650cea2628b6f53c3d53a20fa1aeec3b1c