etcd support, part 1
etcd is used to

- identify agents and physnets attached to those agents;
- pass bind requests to the agents
- confirm that the agents have followed through on the bind request
- store the information required for an agent that's lost a VPP, or a
  VPP that's lost an agent, to get back into sync

This is not quite complete (TODOs abound) but it's working with
non-restarted servers and agents.  To be complete it would require
resyncing in the agent on a restart (and I would recommend killing the
agent if VPP restarts) and a persistent queue - a database table - that's populated in the mechdriver precommit.

The design is predicated on the assumption that etcd is persistent.
Make sure you're running a redundant cluster.

Change-Id: I80f43aa2b5306247eafb84af0e7623c218efde27