Reverse the polarity of the Neutron flow
Previously VPP was server and QEMU client.  However, this seems not to
work for reconnects - current versions of VPP stop listening for new
connections on the server port after a client has connected the first
time.  QEMU is happier with this, so we the code in Neutron tells Nova
to make it the server and the agent makes VPP the client.  Thus is the
Neutron flow reversed in polarity.

The server/clientness of the VPP vhost_user helper function has been
preserved and extended in case we need it back.  The config options
for setting qemu user and group for the agent have been removed, however.

Change-Id: I44a8cd46eff11069635d59b72e7182b485575b08