Recovers etcd watch index on receiving EtcdEventIndexCleared Exception.
Etcd only keeps a history of 1000 events. So if the driver or agent
misses all the 1000 events, which happens when etcd is busy, it recovers
the current state of the watching key space. This is implemented using
an EtcdHelper class in the utils module (new). This module shares common
etcd helper code between the agent and the driver. The eventIndex is
recovered by issuing a read against the keyspace and then starting the
keyspace watch from etcd_index+1.

Change-Id: I6e7cfd7689efc07288bd9613dcc6ad069ada5e7d