Recovers etcd watch keyspace index and updates the driver physnet data structure
Etcd only keeps a history of 1000 events. So if the driver or agent
misses all the 1000 events, which happens when a lot of data is written
to etcd, it is necessary to recover the watching keyspace.
The driver and agent do so by issuing a read against the keyspace
and then re-starting the watch from etcd_index+1. Also, the driver checks
for physnets advertised by the agents when a watch event is received and
updates the physical_networks data structure. This patch does not attempt
to clear old states and sync new port states.

Change-Id: I6e7cfd7689efc07288bd9613dcc6ad069ada5e7d