Upon Agent restart, restart VPP, clear etcd states and bind existing ports
When the agent is restarted, it restarts VPP and clears the old state.
Then it fetches any existing ports and binds the port with VPP and
updates the new state in etcd. The etcd waitIndex a.k.a tick is set
to the correct index so watch can start from etcd-Index + 1

Change-Id: I114e30811559e35a6c0110c2c00c44133bd8de24