Upon Agent restart, restart VPP, clear existing port binding state and bind ports
When the agent is restarted, it restarts VPP and clears the old port
binding state returned to the server from the agent.This assumes that input
data is in etcd, and that the VPP config state and the etcd port binding
information are outputs from the process. On a restart, we don't know whether
we completely propagated all incoming state to VPP and all VPP state to the
port bindings, so we reset both and re-transfer the states. The reset will
cause a data plane disconnect so this needs revisiting to take
account of existing VPP state in the future.  Today we can't add annotations
on VPP objects so that's not practical.

Change-Id: I114e30811559e35a6c0110c2c00c44133bd8de24