Properly set etcd_port
etcd_port was being set multiple times, as $ETCD_USER
and as $ETCD_PASS. I believe this was in error, but more
importantly, it was causing etcd_port to be set a blank
value in the ml2_conf.ini file and leading to neutron failing
to start due to a key error.

This should resolve that.

Change-Id: I68ea27009def89b2501885b1eb9942423e1d6013