Catch the etcd client ReadTimeoutError exception
When the etcd watch times out, the client throws a ReadTimeoutError
exception and logs a stack trace generating an error log on the
console. The primary reason for this error is due to incorrect
logging handlers used by the etcd client library. A complete fix
would require an update to the etcd Client library code. This
patch catches this exception and passes it just as we should
be doing.

Change-Id: I6d6acff50f585f83add775646db5a0d62c7a8df3