Reformulate the etcd read/watch loop as common code
We've already found several interesting ways to add race conditions
to the watch loop, and additionally resyncs are a little hard to
grasp; pull the code out to a common location so that fixes benefit
all watchers and new users don't make all the rookie mistakes.

This includes fixes to address the race window between startup reads
and the first watch.

Change-Id: I55846b346644bb33179ed77979409d38c93fd6a5