Make local etcd worker threads are start after Neutron forks
This is a workaround solution (it's not to be trusted and not
ready for production) to make sure that etcd worker threads
are created per-process.  If they're created before Neutron
forks its workers, all the background threads are waiting on the
same long running etcd query and this causes problems for all
but one of them in updating their physnet cache.

Change-Id: Iec722c58b13305722f9739eea87d8e402043024c