Remove physnet caching mechanism
For unclear reasons, the list of physical networks is not
being populated correctly on Neutron startup, and bindings
fail because the physical network is not found.

As the cache is read only when a port is bound (typically at a VM
creation), and as etcd is meant to be scalable (ie: cluster mode),
we can avoid caching physical network state.  In the long run we
should find a means to avoid remote calls during the client's REST
call, but for now this is an improvement on current behaviour.

Change-Id: I1cc028d0082cda0f5a4761f29f1bae7c7b0a9dc9
Signed-off-by: Christophe Fontaine <christophe.fontaine@qosmos.com>