(WIP) L3 router service plugin for networking-vpp
This is the first iteration of a L3 router service plugin
for networking-vpp. Currently only the internal gateway feature
is implemented in the driver.

WIP: this is part-changed to monitor all the change events (as
PRECOMMIT_*) to fill the journal within the same transaction
and to output the full router config during a change.  However,
it does not work as coded - it doesn't pick up on newly added
routers when it's called in PRECOMMIT_CREATE, the .flush() call
apparently isn't pushing out changes to the DB so that the
.sync_*() calls can pick them up.  The agent also needs changes
(and note the etcd format - as given, a single key with all
interfaces - is not finalised yet - and a different format
may suit better).

Change-Id: Ib7be9a0d396987bfd9a7b353fad337d5c498445f