security group feature implementation.
This is a work in progress for implementing the security
groups feature in VPP. At this stage, the server side portion
of the feature has been implemented. This includes the below
code changes:
1) When security-group rules are
updated, the server receives notifications from neutron
and updates the secgroup in etcd at:
/networking-vpp/secgroups/<security-group-id>
2) Journaling has been updated to include the security group
and allowed-address-pair data.
3) Convert security group rules into a model suitable for
VPP ACLs.

Change-Id: I43a5adb0f94d34a713c3b5c44a3d26f6959d16fd