Liberty compatibility
For those that need it, this will work with Liberty but in one or two
cases uses more modern parts of the Neutron interface.

This patches in the additional functions to the Neutron interface to
make that possible.  It's not terribly sanitary, but the compat layer
is designed not to make changes where they aren't called for.

Change-Id: I33e53c7f3e0c70f2dda9f06a3ce469bad8c8416c