VPP notification handling
Generic implementation to handle events from VPP.

Thanks to this, the agent will now wait for the interface connect
event (ie: VPP is connected to the socket created by QEMU) before
reporting to etcd the connection is done.

We have three layers of object in the agent - the etcd listener,
the VPP forwarder code for the agent, and the VPP API wrapper that
atttempts to come up with a more natural VPP interface.  The
callback for vhost user connections has been added to the low
level interface and the processing of that callback into a
connection event has been added to the agent's forwarder code.

Change-Id: I923fd1fb272c30b35625322f275edb53193c2d77
Co-Authored-By: Ian Wells <iawells@cisco.com>
Closes-Bug: 1647872