Test script to sanity check etcd content
At the moment this is quite limited and only sanity-checks the ports
against the Neutron database.  It will only pass if the two databases
are in perfect sync, which is not a guarantee (etcd follows Neutron
but doesn't always keep up).  And it's looking for etcd on localhost,
and the openstack admin credentials in the environment.

Change-Id: Icd1e7b126680362d5d8fee447cf772f3697a2c0a