Rework interface-get code
We had many layers of get_interface code and no consistency on whether
they returned a name, interface or an object from the VPP interface
(the last probably something we shouldn't let escape from vpp.py so
that we are insulated from VPP API changes).

This change reworks those calls so that they're consistent and named
meaningfully.

Change-Id: I844be16ddb18e3afc71e669f527c2099b7aef984