Use ifidx of uplink interface as bridge domain ID
This means that if we're looking for a bridge domain (in case we're
in resync) we can predict the ID of the domain.  That means, in turn,
that we know if we have BDs that need deleting (because interfaces
that are tagged as ours will have BDs with the same ID and these can
be removed) and crashes where the interface has been created and the
BD hasn't can be detected and easily resolved.

Change-Id: I4ce87a80f101a3f19b8334fe4e8ac1fad15ad122