Make the secgroup code use the journal
The secgroup code as it currently is writes directly to etcd.  We
risk losing changes if the Neutron process terminates between the DB
update and the etcd write.  The journal is here to avoid this by
ensuring all etcd updates are logged in the DB atomically with the
transaction changing the Neutron API objects; we write all changes
via the journal.

The change involves using the precommit (during the transaction after
the API tables have changed) and postcommit (after-commit) hooks to
store the changes required in the journal and wake the forwarder
thread, mirroring the pattern used on the port updates.

Change-Id: Id6aa85807de20920f595a7a6c7d50239ec4adf55