Do something meaningful when physnets are misconfigured
Physnets can point to interfaces that are not available in VPP.

A bug means that this is not identified, and this has been fixed.

A second problem means that were this identified, binding still
proceeds even though the physnet interface is unavailable, which
fails with a cryptic backtrace.  Log meaningful errors and skip
the binding (which will remain in desired state but be ignored
until vpp-agent restarts, eventually timing out with an error
when the caller doesn't see the bind completion).  Resync, after
fixing the config, will get the system into a working state.

Change-Id: Ib7f4b878bb558fb63104599afd7cccaa7168312c