Resync with vpp on agent restart and when losing history
By marking key constructs of the agent when they are created, we can
identify how close to the correct VPP configuration we have at a
restart - adding, changing and cleaning up constructs as required.

By identifying the differences between the key space and the
current port cache, we can work out what changed when we have a gap in
history.

Change-Id: Id34b51c0a550b1d21396beed23d8a9be099fe7ae
Closes-Bug: #1647871
Co-Authored-By: Christophe Fontaine <christophe.fontaine@qosmos.com>