Remove status returns from VPP to make two functions consistent
These two functions have been recently added to return their
success value.  Added a TODO to note the intended behaviour of
problems with VPP (throw an error - which we should catch and use
to indicate that VPP should be terminated) - and removed the
status codes and the use of those codes, to keep the interface
consistent.

Change-Id: Ia2ac565c693082e2c3c58ac9923ba5d0ce748487