Refine VPP interface in vpp.py
Based on the statement of aims description now at the top of the
class, clean up vpp.py.

Clean up return values so that they are returned in a form that does
not relate to VPP and that comes back in a sane format.  Move helper
functions that fix IP address and MAC address formats onto the VPP
class.

Remove some of the 'here is the result of this call' debugging.

Also, reorganise the class to get the ACL and macip_acl functions to
sit together.

This change also alters the current form of the ACL inputs, which
are now converted from a usable form (unpacked macs and IPs, masks
guessed where unsupplied) to the form VPP expects as they are handed
to the function and now no longer at risk of breaking if the
VPP interface changes.

Change-Id: I958b10a75b6fa463d356efe3a266a347d51374b8