Refine VPP interface in vpp.py
Based on the statement of aims description now at the top of the
class, clean up vpp.py.

Move helper functions that convert IP address and MAC address formats
into vpp.py.  Ensure VPP object uses formats that are Python-friendly
and does not leak VPP data structures and abstractions.  This avoids
rewrite of the higher level control code in the event that the VPP
call (based on the VPP message format) changes.

Also, reorganise the class to get various related functions to
sit together.

This change specifically alters the current form of the ACL inputs, which
are now converted from a friendly form (unpacked macs and IPs, masks
guessed where unsupplied) to the form VPP expects as they are handed
to the function and now no longer at risk of breaking if the
VPP interface changes.

Change-Id: I958b10a75b6fa463d356efe3a266a347d51374b8