Update for idempotency
As a preparatory work for resync, we need these methods to have a
consistent behaviour if they are called multiple times with the same
parameters.  We also need to be able to re-find ports to enable this
idempotency.

Methods are renamed from create_* to ensure_* when they are idempotent
as they may not be creating all of the constructs they use in VPP.

Methods have also been renamed to make it clear when they operate
on VPP and when they operate on internal datastructures after
VPP has been configured.  This means that we can reuse those methods
when resyncing, even if the local datastructures are not yet
populated.

Change-Id: Ibb30bc5a4e486c5d43b75feb21b1f85d25f97757
Partial-Bug: #1647871
Signed-off-by: Christophe Fontaine <christophefontaine@mac.com>
Co-Authored-By: Ian Wells <iawells@cisco.com>