(WIP) Add configuration for the command RX queue len of VPP
This cannot yet be tested because the patch has not been applied
to VPP 17.01.

This allows a config item to be used to set the VPP RX queue length.
At the moment it allows us to avoid a bug that appears to be triggered
by short command queue lengths, but it also means that we can avoid
blocking the client when the command queue backs up by increasing its size.

The default setting is 'use the command queue size VPP chooses for its
default.'

Change-Id: Ief6badb6af980907aa734e8aaab65fdc084491ae