Add configuration for the command RX queue len of VPP
This allows a config item to be used to set the VPP RX queue length.
At the moment it allows us to avoid a bug that appears to be triggered
by short command queue lengths, but it also means that we can avoid
blocking the client when the command queue backs up by increasing its size.

The default setting is 'use the command queue size VPP chooses for its
default.'  This means that this will work with versions (at least up to
17.01rc) that don't support this option to VPP jus so long as you don't
set it in your config.

Change-Id: Ief6badb6af980907aa734e8aaab65fdc084491ae