Remove callback activity from within callback
Rather than do VPP work within the callback from the API library,
run activity in a background thread.

This is intended to allow the callback to return as soon as possible.
However, the yield point on semaphore change may prevent that; I'm
not sure this is robust, yet.  However, the other thread working on
queue data will hold the lock for a minimal quantity of time.

Note that additional cleanups have been applied to the event
data manipulation to turn a loop into a dict lookup.

Change-Id: I3ceae2529bb7074ea0706c1ff31281c9fa4d8e8c