Let VPP use native threads
VPP uses a Python thread outside of the eventlet greenpool for its
callbacks.  Using sync primitives from eventlet does not work in this
instance, so we need to instead use threading primitives where
VPP is concerned.  Don't monkey-patch threading so that it can use
the threading-native functions internally, and stick with the threading
library for functions used in its callback.

Locking on the Lock() will stall all eventlets.  This means that calls
to VPP (inside the vpp_papi library) and the lock in the callback
will stall the whole process.  This is acceptable as VPP calls are brief.

Change-Id: I1f227f9ba7d9568bf1cbbca0739aada5ea9a86f9