Resync with vpp on agent restart
This addresses both reasons why resync will happen - the cold
case where the agent has died and requires a restart, and the
warm case where VPP's state is completely known and understood
but Neutron's requests have changed and we no longer have
incremental change notifications.

The port resync is done in this order:

resync-start:
If cold:
  - Read VPP actual configuration and etcd desired configuration
  - Delete anything that lies outside of the sane config options

do_work:
- remove any interfaces that are bound differently than requested
  (using rediscovered or local data as appropriate)
- bind all interfaces (also repopulating /state and causing bind
  notifications and repopulating local DS's if cold)
- correct security groups as required

As bind() is idempotent on VPP, this recycles any elements that
were previously created in an operation that did not complete.
It is also VM-friendly, and leaves interfaces connected where
they are already in the right state.

resync_end:
- remove interfaces that Neutron isn't currently requesting
- (TODO) remove VPP network structures that are no longer in use

Other methods are added in vpp.py to support the resync.

Closes-Bug: #1647871

Change-Id: I655a1532164c80a91ccb6affe15b961751ffa14d
Signed-off-by: Christophe Fontaine <christophe.fontaine@qosmos.com>
Co-authored-by: Christophe Fontaine <christophe.fontaine@qosmos.com>