Resolve too many threads running on the etcd and DB journal
Closes-bug: #1644371

Implement an election algorithm to elect two threads
as masters for doing work - one for DB journal (forward workers)
and the other for etcd (reverse workers). The threads that are
not elected as master go to sleep and check if the master is
alive periodically. If they cannot detect a master, a new master
is elected to do the work.

Change-Id: I13a4dbe4e58c8156231acf7fab2ca293f7b4e4cb