Add new etcd watcher model
This addes a new subclass to EtcdWatcher that attempts to analyse
changes in etcd itself, regardless of whether a watch or resync
happened, and will then pass on the calls to its listening function
accordingly.

This has the advantage that resyncs because of history problems,
syncs because of initial setup and watches are all reduced to the same
problem.

This makes the system more susceptible to port-bind-before-secgroup
problems.

Change-Id: Ib49a5efde84171a387d2ba468bb16999a42db56f