Register before_delete event for liberty security group deleting
In Liberty, with its AFTER_DELETE hook, there's no way to get the
security group ID to refresh because the information has been lost from
the DB and is not otherwise available.  Add a memory, populated in
BEFORE_DELETE, that will provide the secgroup rule to AFTER_DELETE.

Additionally, clean up the deleted-rules ignorer to clean up rules just
once.

Change-Id: I2793350db0f01d4c9fea310a9c0ed9830c44bf3d
Closes-Bug: 1667860