Disable l2 tag rewrite for vhost interface
vpp will cache last l2 tag rewrite status of a deleted port on
the same index, so there's chance that a new created vhost
interface is on a index that was a vlan subinterface. Here
force to disable l2 tag rewrite when creating a vhost interface,
in order to ensure all vhost interface are not rewriting l2 tag.

Change-Id: I6dbea77f6c2057deeaf688b0a9cffed1b0a2b313
Closes-Bug: 1667849