(WIP) Use separate clients for each of the eventlets/threads
We use multiple eventlets in the system, but they share clients.
We've seen this leading to one call getting the answer for another,
and it's possibly also the cause of Neutron hanging (theoretically,
because it's in a long etcd watch when it sends another etcd request).

Change the code to ensure that background threads use their own etcd.

This may provoke bind-before-secgroup races, which need fixing
independently.

Change-Id: I0487912e9d5115a70a5a4277abf6c54105755a15