Update binding notification code
It seems that we beat Nova to the punch, sometimes noticing that a
port is bound before Nova thinks it's possible.  In this case,
we send an event to Nova to signal bind completion, but it isn't
yet waiting for it.  Notify a second time, 4s later, as a temporary
fix until we can find a better solution (in Neutron or Nova as the
case may be).

Additionally, use the memory regions as the indicator of binding.
We aren't certain of what sock_error would be on an unbound port,
but we know the port is definitely connected if memory regions
are present.

Change-Id: I512ca70789ed5628e10a819a39af5793feb6a505