Remove secgroup push workaround
We pushed all secgroups before we bound a port, for two reasons:

1. there was a bug in writing security groups such that they weren't
always output on creation and not present when the port was bound.
This has long been addressed
2. there was no link between the secgroup thread, and, while no
guarantee, this made it considerably more likely secgroups were
present before a port was bound.  Async secgroup waiting fixes
this.

With the above addressed, the workaround can be removed.

Change-Id: Ic5c8e8c46774564d4b28aa844df41b187144d0dd