Add provisioning blocks support
For modern versions of OpenStack that support it, rather than putting
the port to ACTIVE in the mechanism driver, it is necessary to place
a provisioning block there and release it when provisioning finishes,
allowing ML2 itself to determine that all jobs are complete and that
the port should be made active.  Add support for those versions.

Removed some other hacks related to this that weren't working as they
should have (e.g. checking memory_regions doesn't work as a notification
source, and sock_errno wasn't the source of the problem).

Change-Id: I248fbca1d00f313075ad34f7cd1baceb4742786f