Resyncing on agent restart
Resyncing will:

- clean up uplinks that are no longer as they were in the
  configuration
- remove all (tagged) interfaces if they're not known in etcd
  (we can leak interfaces very slowly since if there's a crash
  between creating the interface and tagging it we won't be able
  to tell we have work)
- remove unused bridges and bridges where physnet config
  has changed
- rebind all interfaces that are known in etcd (creating new bridges
  for config-changed items)
- rebind all the VM interfaces, recreating internal datastructures

This is not guaranteed to work for GPE networks in this revision.

This does not attempt to delete networks directly; there should be
no unused networks on a restart, but poor timing can lead to empty
bridges and that's a potential leak.  Used networks will be
re-identified in the rebind phase.

Change-Id: I3ff5d46f7b61065d530ed79cf00cee3627a00bb7
Closes-Bug: 1647871