Deal with uplink, upstream and physnet interfaces
We confusingly use 'upstream' to mean either uplink or
physnet interface and this is not making the code readable.
The use of 'upstream' is made clear (i.e. removed in favour of
the correct term.

The physnet is added to uplink tags.  Physnet interfaces are
independently tagged when used.  (Where a physnet interface is
also an uplink, the physnet tag wins.)

Change-Id: Ica50e56b9d6b5979dd5d7c45569a24f2ccae4dd5