Don't delete subinterfaces when they don't exist
This is not the end of this line of fixing, but it should prevent
the deleting of a subinterface automatically when no subinterface is
in use for a network.

This may need further work to make it more robust in cases where e.g.
a restart caused a bridge but not a subnet to be created before
crashing out.

Change-Id: Ibfbd98696097da6b464f77f2110988dcd4eda00a