Isolate vpp_etcd_journal table for update lock to a single row
Earlier identifying the earliest entry to the vpp_etcd_journal table and
locking it for update, so that we can process it and delete it was done
in a single query. We suspect, this might be accidentally causing a
lock on all the entries in the table (as part of the read) and causing
side effects when other threads are waiting for this table to be unlocked

The patch aims to split this into two parts.
(a) A normal select query to identify the oldest row and
(b) then a 'select for update' query on that row alone.
When multiple threads are entering this block, all the others who got the same
row id, will wait till the one with the lock releases it . It will query and find
that the row is gone and will starting another read.

Change-Id: I39bced823d7e1a7c6b38ab5490117ffced398741
Closes-bug:#1691813