(WIP) JWT signing and verification library for k/v stores
This is a first pass at extracting JB's code into a signer and verifier
for key-value stores.  I've used the base code, Pythonified it a bit
and tried to make it as much a standalone module as possible - to the
point that it's not logging data any more, and doesn't read config
directly, so is low on dependencies.

It should be suitable for gluing into a shim object installed onto
etcd_client and replacing our direct calls to the python-etcd client.

Change-Id: Iaa1a70a7e87b935b6ff48e2b1f27784ed27ecc97