Support remote-group-id in security-group rules
Prior to this patch, the remote-group-id attribute in security-group
was ignored by networking-vpp. This patch adds support for the remote-group-id
attribute by using an agent thread to watch for changes in security-group
to port associations. The agent then computes VPP ACLs by taking a product
of the port IP addresses in the remote-group and other rule attributes
such as protocol and TCP/UDP port.

Closes-Bug: #1656471
Change-Id: If68ab2c4cba179ec181d10a9d80ca40d78c76a0c