Timeout etcd_update
Adding a configurable eventlet timeout on the etcd_update function
in the vpp journal forward worker. This will prevent the master thread
from being hijacked by a slow or non responsive etcd.

Also adding a master time slice to determine the period for which the
master thread remains master (it sets the internal PARANOIA_TME
used in master election). This is configurable and set to 120 seconds
to handle DB timeout scenarios.

Change-Id: I480a4ec44b571c24fcf42520220de0439743ef81
Closes-bug:#1694723