(WIP) Per-thread etcd clients
Automate the use of etcd clients so that each thread gets its own
without conscious thought.  This should reduce the odds of two
eventlets going for the same etcd client and interleaving their
queries (as etcd connections are not eventlet-safe).

The programming model is not perfect, since we do continue to
save the etcd_client for a thread in various data structures, but
it's noticeably improved.

Change-Id: I89c053a3cef58e01836b44afaf2c1453ae052525