Add configurability to etcd running
etcd3 is now available as a devstack service.  Do not install or
run etcd if devstack thinks it's going to (particularly important
since devstack doesn't use the distro-provided service).

Make running etcd command-line configurable, as it's perfectly
reasonable to use an existing etcd on another host or hosts.

Make cleaning etcd configurable, as in a multi-machine environment
you probably don't want devstack emptying out the etcd keys unless
it's also installing the DB afresh (which won't be true when there
are compute nodes or if there are multiple controllers).

Change-Id: Ia8e5c6e31843bb1ddf8102e208d734deb2706faa
Closes-Bug: 1707821