Delete GPE remote keys from etcd when ports are unbound
If the vpp-agent is not active at the time GPE ports are unbound,
the remote GPE keys in the gpe key space are not removed. If the
the vpp-agent is started at a later time, it tries to create stale
remote mappings for these already unbound ports. This patch resolves
this issue.

Closes-Bug: 1710990
Change-Id: I741015622c8b97d19f0e4532e799b09ed352aab7