Adding eventlet worker to ensure tap interfaces in the right bridge
When DHCP or L3 agent crashed or restarted, the corresponding tap
interfaces created in the kernel will also be removed and recreated. In
that case, VPP has no knowledge of the tap interface being vanished and
respawned, so those tap interfaces won't be plugged into the right Linux
bridge after recovery. As a result, the plumbing between VPP and DHCP/L3
agent is broken.

The patch will add an eventlet worker to watch for interface create and
delete events from kernel, and ensure those tap interfaces are in the
right bridge.

Change-Id: I796087f7917da8a838d39e9821b34cf58473d7cf