(WIP) Shim for etcd data storage
Where we are using etcd to store data for communication between server
and agent, use a shim.  The shim does the serialisation and
deserialisation of the data.  This will allow us, in future, to
further manipulate the date (for instance, by signing) by extending
the writing shim.

Not all data is written via the shim - elections and other more
basic components are left untouched as of now.

Change-Id: I9fa13df0bb106fa8ea82da757747d58d67ab1a11