(WIP) Updates to the router service plugin
Update the router service plugin to manage router states in etcd
as opposed to the command based design. Also, a router interface
can now be added to a GPE subnet.
The router service plugin will write a state key to
the below etcd directory whenever the router is updated:
/networking-vpp/nodes/<network-node-name>/routers/<router-id>.
The state is written for the below router update operations:
1) Setting an external gateway
2) Adding a router port on a subnet
3) Adding the router to a subnet
A delete operation deletes the corresponding state fron etcd.

Change-Id: I6af6a4ef959aafb4f3105029fc7428db349b0017