L3 router service plugin enhancement
The following enhancements have been made to the L3 router
service plugin in networking-vpp.
1. Updated the L3 service plugin to manage router and
floating-ip states in etcd as opposed to the previous command
based design.
2. Added support for re-syncing both the router and FIP states
when the vpp-agent is restarted.
3. Added support for attaching a router interface to a
VXLAN GPE segment.
4. Added support for external gateway connectivity by exporting
routes between the tenant VRFs and the default VRF0.
5. Added support for attaching any gateway port on the network to
the VPP router. i.e. we now support the 'openstack router add
port' operation besides the 'add subnet'.

The router service plugin will write a state key to
the below etcd directory whenever the router is updated:
/networking-vpp/nodes/<network-node-name>/<router-id>/<port-id>
The state is written for the below router update operations:
1) Setting an external gateway
2) Adding a router port on a subnet
3) Adding the router to a subnet
4) Associating a floating IP address
A delete operation removes the corresponding state key from
etcd.

Change-Id: I6af6a4ef959aafb4f3105029fc7428db349b0017