(WIP) Update the router service plugin to support states in etcd
Update the router service plugin to manage router states in etcd
as opposed to the previous command based design.
This patch also includes support for
1) Adding a router interface to a vxlan GPE segment.
2) Creating a neutron port on a network and binding it to the
router.
The router service plugin will write a state key to
the below etcd directory whenever the router is updated:
/networking-vpp/nodes/<network-node-name>/<router-id>/<port-id>
The state is written for the below router update operations:
1) Setting an external gateway
2) Adding a router port on a subnet
3) Adding the router to a subnet
A delete operation removes the corresponding state from etcd.

Change-Id: I6af6a4ef959aafb4f3105029fc7428db349b0017