Do not add external tap into bridges when binding ports
External TAP interfaces created by DHCP or L3 agent will be added to
corresponding Linux Bridge by vpp-agent to talk to VPP. During a regular
port binding process, there are two code paths calling the same function
for adding the interface to the Linux Bridge, which may potentially
cause a race condition and a non-harmful traceback in the log.

The fix will eliminate the non-harmful traceback in the log.

Change-Id: Ie32704bd397d00a1272ec7a8b4e0ae1426151f2c