Fix unbind error when deleting an external network.
When an external subnet is deleted, if there are multiple DHCP agents,
neutron places the DHCP server's port in a binding-failed state on
compute hosts. During network deletion, the port_delete_precommit call
fails in networking-vpp as it assumes that the port has a valid binding
level. This patch resolves this issue.

Change-Id: I2f5e37da922c47c280828b20f39c127038b5e68d