Enable live migration support
Nova live migration will not currently succeed in ML2/VPP plugin enabled.
(with vif_plugging_fatal=True). This is because as part of live migration, Nova
starts bringing up VM in the new host, but no messages are sent to Neutron,
atleast initially. Nova expects the L2 agent to plump the VMs VIF and set port
status to active, after which it updates Neutron about the new binding host.
But this is based on the assumption that the L2 agent will bind the port without
any signal from ML2  core plugin. But only reference ML2 agents do that, because
they are actively scanning the host and can call back to ML2 core about a new
interface and its binding porperties when they detect one.

In the case of VPP, we are driven by the ML2 plugin for all port bindings and
will be never informed of a new port to be bound. The solution proposed is to
watch out for vhost socket creation events and close the vhost socket by
creating VPP interface for it so that Nova can bringup the VM successfully.

We add a general filemonitoring utility to VPP agent driver which scans the
vhost socket user directory as set by config. When a new file that matches
neutron's port interface pattern is detected, we let vpp create an interface for
it if not already. This will let Nova migrate the VM and later update the port
binding successfully.

Other notes:
- We only look for inotify CREATE and DELETE events and ignore others.
- VPP interface mac address generated from the neutron port itself.

ToDo:
A better way of handling event overflow is needed. Added a tracked files index,
but is not fully developed.

Change-Id: Id372a9863cb41fc3d13f54c77da5032607990b05