Resolve a bug in the security group rule with integer IP Protocol values
When a SG rule is created in Neutron with an integer IP protocol value,
this rule does get configured and the vpp-agent throws an error.
Neutron sets the protocol value to a string which was not being correctly
converted to an integer. Resolved the issue by updating the mech-vpp
driver code to convert protocol values from string to int.

Change-Id: I6ad1278a7364df9e30b7351adad9579e2369878e
Closes-Bug: #1737578