Resolve security group rule with integer protocol values
When a SG rule is created in Neutron with an integer protocol value,
this rule does get configured on the VPP and the vpp-agent
throws an error. Neutron sets the protocol value to a string which
was not being correctly converted to an integer. Resolved the issue by
updating the mech-vpp driver code to recognize empty port ranges and
correctly converting from string to int.

Change-Id: I6ad1278a7364df9e30b7351adad9579e2369878e