Change loading of SignedEtcdJSONWriter
In advance of making signing a loadable extension, move the signing
to a separate file that is loaded only when signing is configured.  This
has the additional effect of not loading the crypto libraries in cases
where signing is not used.

As an aside, remove the 'encoder' argument that was set to
EtcdJSONWriter but never actually used.

Change-Id: I090628b98a2436504be7fa5a8ee8612523543375