Add plugin code manager for agent and mechanism driver plugins
Standalone agent plugins are intended to add functionality to the agent
in response to additional data from e.g. service drivers associated with
the mechanism driver.

Driver plugins are best suited for non-API extensions to the system
(like etcd value signing).

Hooks are provided to the extensions from other extensions and the
base class to allow them to be called on certain events.

Change-Id: I89c6c4f222abf5c3eca8d8683dbcd513a9dd91d6