Make L3 HA configurable
The L3 HA feature is enabled by default. This leads to undesirable
consequences for the simple non-ha deployments consisting of just a
single network/controller node.

Adding a new config option, enable_l3_ha, which would need to be
set to True explicitly in the configuration file to enable the
feature.

Also, added log messages to indicate that L3 HA feature is turned
ON.

Change-Id: Iaf8699adc76bc15fb29fa0d26d3da4051547cbf2