Convert lock to eventlet to prevent deadlocks
inotify and the etcd watcher eventlets compete to use
ensure_interface_on_host, and to keep the datastructures safe it
requires an entry/exit lock.  However, the previous oslo_concurrency
implementation incorrectly guessed that a threading lock was required,
which (in eventlet, which is coroutine based and runs in one thread)
deadlocks.  Convert the lock.

Change-Id: I052a74c8a3b02fd233cb089123d3bb4fae576e60