Avoid tox-install.sh
Instead of using tox-install.sh, we can install neutron
now directly from pypi - or use the current branch in CI system using
tox-siblings.

Update tox.ini for this change and add neutron to tox for now
(there are probably better places, but requirements.txt cannot be used
because it's not a requirement for the agent).

Change-Id: I2b9817923dd540b47c7341ae4623f4512aa9e14b